requirements.txt
numpy>=1.21.0
opencv-python-headless>=4.5.0
onnxruntime>=1.22.0
Pillow>=9.0.0